Max#5
Complete lobby logic
Sync lobby connection
Sync lobby View

Correct player identification on game scene (human/monster)
