+++ HideInTheLight/Assets/Develop/Prefabs/GamePlayer.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0f177ee09f6c10c4f8ce806aad923046
+guid: 662ff0aa8e5f13d429d0c2d53dc32ba1
 PrefabImporter:
   externalObjects: {}
   userData:
