+++ HideInTheLight/Assets/Develop/Scripts/Bootstrap/Old/LobbyManager.cs
@@ -1,4 +1,5 @@
 using Develop.Scripts.Core.Lobby.Old;
+using Develop.Scripts.Core.Old;
 using Mirror;
 using System.Linq;
 using UnityEngine;
