+++ HideInTheLight/Assets/Develop/Scripts/Core/Lobby/LobbyModel.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5f533d609238c024ebbf2096042bda9b
+guid: d1c7e121af115d644bda84cd200d8f7a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
