+++ HideInTheLight/Assets/Develop/Scripts/Core/Lobby/LobbyView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c73002b0cc80d344d9be2cb26da05959
+guid: 5f533d609238c024ebbf2096042bda9b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
