+++ HideInTheLight/Assets/Develop/Scripts/Core/Old.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 0f177ee09f6c10c4f8ce806aad923046
-PrefabImporter:
+guid: de62e4c46fc4c554c97da35fd10680ca
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
